Rename files and move software options to options.
