+++ alphastats/gui/pages/02_import_data.py
@@ -8,7 +8,7 @@ try:
         get_sample_names_from_software_file,
         read_uploaded_file_into_df,
     )
-    from alphastats.gui.utils.software_options import software_options
+    from alphastats.gui.utils.options import software_options
     from alphastats.gui.utils.ui_helper import sidebar_info
     from alphastats.loader.MaxQuantLoader import MaxQuantLoader
 
@@ -18,7 +18,7 @@ except ModuleNotFoundError:
         get_sample_names_from_software_file,
         read_uploaded_file_into_df,
     )
-    from utils.software_options import software_options
+    from utils.options import software_options
     from alphastats import MaxQuantLoader
     from alphastats import DataSet
